+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 8.0.0 / YYYY-MM-DD
+# 8.0.0 / 2018-10-16
 
 ## BREAKING CHANGES